# colexec: convert hashjoiner to use execgen directives