… pkg/sql/colexec/execgen/cmd/execgen/main.go
… pkg/sql/colexec/hashjoiner_tmpl.go