- Add "mid-stack inlining" to `// execgen:inline`
- Use execgen directives in `hashjoiner_tmpl` instead of text/template.

Hashjoiner_tmpl is now text/template free! It's somehow also slightly faster, but I'm not sure why:


```
name                                                                   old time/op    new time/op    delta
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=2048-24        397µs ± 1%     385µs ± 0%  -3.14%  (p=0.000 n=9+8)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=262144-24     39.2ms ± 5%    36.7ms ± 2%  -6.36%  (p=0.000 n=10+10)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=4194304-24     1.00s ± 2%     0.97s ± 1%  -2.77%  (p=0.000 n=9+9)

name                                                                   old speed      new speed      delta
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=2048-24      330MB/s ± 1%   341MB/s ± 0%  +3.24%  (p=0.000 n=9+8)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=262144-24    429MB/s ± 5%   457MB/s ± 2%  +6.70%  (p=0.000 n=10+10)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=4194304-24   269MB/s ± 3%   277MB/s ± 1%  +2.84%  (p=0.000 n=9+9)

name                                                                   old alloc/op   new alloc/op   delta
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=2048-24        893kB ± 0%     893kB ± 0%    ~     (p=0.700 n=10+10)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=262144-24     52.3MB ± 0%    52.3MB ± 0%    ~     (p=0.896 n=10+10)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=4194304-24     777MB ± 0%     777MB ± 0%  -0.00%  (p=0.029 n=9+10)

name                                                                   old allocs/op  new allocs/op  delta
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=2048-24          142 ± 0%       142 ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=262144-24      1.29k ± 0%     1.29k ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=4194304-24     16.7k ± 0%     16.7k ± 0%    ~     (p=0.839 n=8+9)
```

Release note: None